Import PR list API implementation to `pr/list` package
Also splits List vs. Search queries into separate methods for better
maintanability.
