+++ api/queries_pr.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/cli/cli/internal/ghinstance"
 	"github.com/cli/cli/internal/ghrepo"
-	"github.com/cli/cli/pkg/githubsearch"
 	"github.com/shurcooL/githubv4"
 	"golang.org/x/sync/errgroup"
 )
@@ -908,173 +907,6 @@ func isBlank(v interface{}) bool {
 	default:
 		return true
 	}
-}
-
-func PullRequestList(client *Client, repo ghrepo.Interface, vars map[string]interface{}, limit int) (*PullRequestAndTotalCount, error) {
-	type prBlock struct {
-		Edges []struct {
-			Node PullRequest
-		}
-		PageInfo struct {
-			HasNextPage bool
-			EndCursor   string
-		}
-		TotalCount int
-		IssueCount int
-	}
-	type response struct {
-		Repository struct {
-			PullRequests prBlock
-		}
-		Search prBlock
-	}
-
-	fragment := `
-	fragment pr on PullRequest {
-		number
-		title
-		state
-		url
-		headRefName
-		headRepositoryOwner {
-			login
-		}
-		isCrossRepository
-		isDraft
-	}
-	`
-
-	// If assignee wasn't specified, use `Repository.pullRequest` for ability to
-	// query by multiple labels
-	query := fragment + `
-    query PullRequestList(
-		$owner: String!,
-		$repo: String!,
-		$limit: Int!,
-		$endCursor: String,
-		$baseBranch: String,
-		$labels: [String!],
-		$state: [PullRequestState!] = OPEN
-	) {
-      repository(owner: $owner, name: $repo) {
-        pullRequests(
-			states: $state,
-			baseRefName: $baseBranch,
-			labels: $labels,
-			first: $limit,
-			after: $endCursor,
-			orderBy: {field: CREATED_AT, direction: DESC}
-			) {
-				totalCount
-				edges {
-					node {
-						...pr
-					}
-				}
-				pageInfo {
-					hasNextPage
-					endCursor
-				}
-			}
-		}
-	}`
-
-	var check = make(map[int]struct{})
-	var prs []PullRequest
-	pageLimit := min(limit, 100)
-	variables := map[string]interface{}{}
-	res := PullRequestAndTotalCount{}
-
-	// If assignee was specified, use the `search` API rather than
-	// `Repository.pullRequests`, but this mode doesn't support multiple labels
-	if assignee, ok := vars["assignee"].(string); ok {
-		query = fragment + `
-		query PullRequestList(
-			$q: String!,
-			$limit: Int!,
-			$endCursor: String,
-		) {
-			search(query: $q, type: ISSUE, first: $limit, after: $endCursor) {
-				issueCount
-				edges {
-					node {
-						...pr
-					}
-				}
-				pageInfo {
-					hasNextPage
-					endCursor
-				}
-			}
-		}`
-		q := githubsearch.NewQuery()
-		q.SetType(githubsearch.PullRequest)
-		q.InRepository(ghrepo.FullName(repo))
-		q.AssignedTo(assignee)
-		q.SortBy(githubsearch.CreatedAt, githubsearch.Desc)
-		if states, ok := vars["state"].([]string); ok && len(states) == 1 {
-			switch states[0] {
-			case "OPEN":
-				q.SetState(githubsearch.Open)
-			case "CLOSED":
-				q.SetState(githubsearch.Closed)
-			case "MERGED":
-				q.SetState(githubsearch.Merged)
-			}
-		}
-		if labels, ok := vars["labels"].([]string); ok && len(labels) > 0 {
-			if len(labels) > 1 {
-				return nil, fmt.Errorf("multiple labels with --assignee are not supported")
-			}
-			q.AddLabel(labels[0])
-		}
-		if baseBranch, ok := vars["baseBranch"].(string); ok {
-			q.SetBaseBranch(baseBranch)
-		}
-		variables["q"] = q.String()
-	} else {
-		variables["owner"] = repo.RepoOwner()
-		variables["repo"] = repo.RepoName()
-		for name, val := range vars {
-			variables[name] = val
-		}
-	}
-loop:
-	for {
-		variables["limit"] = pageLimit
-		var data response
-		err := client.GraphQL(repo.RepoHost(), query, variables, &data)
-		if err != nil {
-			return nil, err
-		}
-		prData := data.Repository.PullRequests
-		res.TotalCount = prData.TotalCount
-		if _, ok := variables["q"]; ok {
-			prData = data.Search
-			res.TotalCount = prData.IssueCount
-		}
-
-		for _, edge := range prData.Edges {
-			if _, exists := check[edge.Node.Number]; exists {
-				continue
-			}
-
-			prs = append(prs, edge.Node)
-			check[edge.Node.Number] = struct{}{}
-			if len(prs) == limit {
-				break loop
-			}
-		}
-
-		if prData.PageInfo.HasNextPage {
-			variables["endCursor"] = prData.PageInfo.EndCursor
-			pageLimit = min(pageLimit, limit-len(prs))
-		} else {
-			break
-		}
-	}
-	res.PullRequests = prs
-	return &res, nil
 }
 
 func PullRequestClose(client *Client, repo ghrepo.Interface, pr *PullRequest) error {
