+++ pkg/cmd/pr/list/fixtures/prListWithDuplicates.json
@@ -2,33 +2,27 @@
   "data": {
     "repository": {
       "pullRequests": {
-        "edges": [
+        "nodes": [
           {
-            "node": {
-              "number": 32,
-              "title": "New feature",
-              "url": "https://github.com/monalisa/hello/pull/32",
-              "headRefName": "feature"
-            }
+            "number": 32,
+            "title": "New feature",
+            "url": "https://github.com/monalisa/hello/pull/32",
+            "headRefName": "feature"
           },
           {
-            "node": {
-              "number": 32,
-              "title": "New feature",
-              "url": "https://github.com/monalisa/hello/pull/32",
-              "headRefName": "feature"
-            }
+            "number": 32,
+            "title": "New feature",
+            "url": "https://github.com/monalisa/hello/pull/32",
+            "headRefName": "feature"
           },
           {
-            "node": {
-              "number": 29,
-              "title": "Fixed bad bug",
-              "url": "https://github.com/monalisa/hello/pull/29",
-              "headRefName": "bug-fix",
-              "isCrossRepository": true,
-              "headRepositoryOwner": {
-                "login": "hubot"
-              }
+            "number": 29,
+            "title": "Fixed bad bug",
+            "url": "https://github.com/monalisa/hello/pull/29",
+            "headRefName": "bug-fix",
+            "isCrossRepository": true,
+            "headRepositoryOwner": {
+              "login": "hubot"
             }
           },
           {
